LayoutTests/platform/mac-monterey/imported/w3c/web-platform-tests/xhr/open-url-redirected-sharedworker-origin-expected.txt


PASS Referer header
FAIL Origin header assert_equals: expected "Origin: http://localhost:8800\n" but got ""
PASS Request URL test
